requirements.txt
streamlit
mediapipe
numpy
Pillow
opencv-python-headless